requirements.txt
discord.py==2.3.2
llama_cpp_python==0.2.27
python-dotenv==1.0.0

